Unify memory tracking by centralizing implementation in utils::mem
